+++ src/icemac/ab/calexport/conftest.py
@@ -8,18 +8,26 @@ import pytest
 
 
 pytest_plugins = (
-    'icemac.addressbook.conftest',
-    'icemac.ab.calendar.conftest'
+    'icemac.addressbook.fixtures',
+    'icemac.ab.calendar.fixtures'
 )
 
 
 # Fixtures to set-up infrastructure which are usable in tests:
 
 
+@pytest.yield_fixture(scope='function')
+def address_book(addressBookConnectionF):
+    """Get the address book with calendar as site."""
+    for address_book in icemac.addressbook.conftest.site(
+            addressBookConnectionF):
+        yield address_book
+
+
 @pytest.fixture('function')
 def browser(browserWsgiAppS):
     """Fixture for testing with zope.testbrowser."""
-    assert icemac.addressbook.conftest.CURRENT_CONNECTION is not None, \
+    assert icemac.addressbook.testing.CURRENT_CONNECTION is not None, \
         "The `browser` fixture needs a database fixture like `address_book`."
     return icemac.ab.calexport.testing.Browser(wsgi_app=browserWsgiAppS)
 
@@ -47,14 +55,46 @@ def MasterDataFieldFactory(FieldFactory):
     return create_masterdata_field
 
 
+# generally usable helper fixtures:
+
+
+@pytest.fixture(scope='function')
+def sitemenu(browser):
+    """Helper fixture to test the selections in the site menu."""
+    return icemac.addressbook.testing.SiteMenu
+
+
 # Infrastructure fixtures
 
 
 @pytest.yield_fixture(scope='session')
-def zcmlS(zcmlS):
+def zcmlS():
     """Load calendar export ZCML on session scope."""
     layer = icemac.addressbook.testing.SecondaryZCMLLayer(
-        'CalendarExport', __name__, icemac.ab.calexport, [zcmlS])
+        'CalendarExport', __name__, icemac.ab.calexport)
     layer.setUp()
     yield layer
     layer.tearDown()
+
+
+@pytest.yield_fixture(scope='session')
+def zodbS(zcmlS):
+    """Create an empty test ZODB."""
+    for zodb in icemac.addressbook.testing.pyTestEmptyZodbFixture():
+        yield zodb
+
+
+@pytest.yield_fixture(scope='session')
+def addressBookS(zcmlS, zodbS):
+    """Create an address book for the session."""
+    for zodb in icemac.addressbook.conftest.pyTestAddressBookFixture(
+            zodbS, 'CalendarS'):
+        yield zodb
+
+
+@pytest.yield_fixture(scope='function')
+def addressBookConnectionF(addressBookS):
+    """Get the connection to the right demo storage."""
+    for connection in icemac.addressbook.conftest.pyTestStackDemoStorage(
+            addressBookS, 'CalendarF'):
+        yield connection